Remove outdated node engine
We recently recieved a message from Heroku informing us that the
node engine we had in package.json was no longer secure. I'm just
going to remove it and allow Heroku and anyone else to grab
the modules for whatever node version they have. There's
nothing particularly complex that needs to be pegged to a
specific version.